add getters and setters for dates and format them
by default, if a user provides a due date but no time, it will go to 12am of that due date. this feels like the expected behaviour. can always change if i am wrong. if user does not provide due date / time using setter, then the object will just not have a due date but will eventually still show up in the list of tasks