write getter and setter (storer) of projects and tasks.
used json.stringify on the key (which is the project object) as, iiuc, you cannot have an object as a key of an object. if we only use the name of the project (a string, for sure) as the key of the object in localstorage, we will need to double up the amount of work, i foresee, if we edit the name of the project. the link between the object of the project (with all of its methods etc) and the name of the project breaks too easily if we only use the project name as the key in localstorage. anyways, i might change this later on once i learn more...