in project factory, the createdAt date is generated when it is created. add getter for that createdAt attribute for project.
because we presume that the user cannot simultaneously create two separate to-do list tasks, created timestamp is sufficiently unique for our purposes. it is a private variable that cannot be edited. it shouldn't be edited at all, in fact, which is why there's no setter method for it either in factory