store projects and their properties (including their tasks) in localStorage, and be able to retrieve them (#2)
* add details of task to the project rather than just the object

this allows the console to be able to actually see each of the values

* add tasks to localStorage via the project, 'My Tasks' using the getter to store the actual task list (which we've made private and only accessible through the getter getTask())

* get 'task' array from localStorage

* need to stringify the tasks in myProject as only strings can be added to localStorage (as it is JSON). we are, instead, trying to add an array of objects.

* separate out createProject factory as a module for easier maintainability of code

* add function to get stored tasks

* add function to store a particular project to localStorage. log tasks that are stored if they exist, otherwise say that there are no tasks yet in the console so user knows

* add comment question because i'm confused. store project to the localStorage

* end main IIFE function by showing all locally stored tasks

* comment out clearing of storage. add comment about bug identified as to do

* move into subfolders specifically for factories

this is to prep for creating a new folder for utils, in which a localstorage module will contain helper functions we are about to write for this branch

* add utility folder in src and localStorageHelpers module with comments for clarity

i might remove these comments later on, but right now, it is unfortunately the only way i can stay sane and focused on the task at hand

* export localstorage helper functions and import them into index.js

* write getter and setter (storer) of projects and tasks.

used json.stringify on the key (which is the project object) as, iiuc, you cannot have an object as a key of an object. if we only use the name of the project (a string, for sure) as the key of the object in localstorage, we will need to double up the amount of work, i foresee, if we edit the name of the project. the link between the object of the project (with all of its methods etc) and the name of the project breaks too easily if we only use the project name as the key in localstorage. anyways, i might change this later on once i learn more...

* ensure correct fu helper functions relating to localstorage are called in index.js

* in project factory, the createdAt date is generated when it is created. add getter for that createdAt attribute for project.

because we presume that the user cannot simultaneously create two separate to-do list tasks, created timestamp is sufficiently unique for our purposes. it is a private variable that cannot be edited. it shouldn't be edited at all, in fact, which is why there's no setter method for it either in factory

* enable creation of project in question flow

* expose tasklist of project as public property

easier to manipulate while i test how to work with it, as well as with localstorage

* check if there's any projects created yet before creating project

* change data structure per updated comment in helpers. store project using method. export and then import into index.js to use on the generic, catch-all 'my tasks' project which only gets created and added to localstorage if there's nothing inside of localstorage

* remove unnecessary boolean check

just write it directly in index.js for now

* add comments

* store newly created project into localStorage

* expose tasks' title and description as public properties

easier to manipulate. these properties also don't have checks. other properties (like priority) has only a set of valid values which needs to be checked. exposing public property would circumvent these

* add warning comment

* expose projects' name property. this doesn't need any checking at the moment and can be manipulated

easier to manipulate for us right now. i might make private in future.

* simplify viewDetails function using destructuring

* simplify viewDetails function using destructuring

* simplify exports i n tasks factory

getters are not used - expose when they need to

* remove returned method that wasn't used anymore in project factory

* create viewdetails method for project factory too

purpose is to add these as the data into localStorage

* store projects and all of its details (id, name, list of tasks) into localstorage. enables adding new projects

* use shorthand for arrow function to get stored project. retrieve data in my tasks project in localstorage based on name

retrieving my unique id is great (requires storing as such too). however, it would mean double the effort when we create a new project. i might revisit this later on

* add alternate way to store a project into localstorage in the project object as well

* Revert "add alternate way to store a project into localstorage in the project object as well"

This reverts commit 56c6cf120719a59ccdb58640f4a467430f8ebded.

* add newly-created task to My Tasks ('catch all' project)

* if user does not want to add a newly-created task to a project, it will be created and stored in the localstorage under 'my tasks', the catch-all project

* simplify naming of methods

* write methods to retrieve all projects as an array, and retrieve everything as an object (key being the project name, value being the various attributes of the project, including the tasks list)

* remove comments

* remove comments