retrieve the actual project objects (not just the names) from get go
this enables us to grab the objects' id from localStorage. i decided not to reconstruct the project object as we're not writing to the project (and thus messing about with a private variable), just grabbing it and sorting the projects. if i want to be absolutely safe, then, in the future, i would create a getter for project id, reconstruct every project we retrieve