store tasks' due dates in format YYYY-MM-DD
if, when displayed, it needs to be reformatted, we will then use date-fn library in the DOM. but, converting it at the Task factory level, storing it, and then changing it back feels cumbersome, especially because no normal user is going to go looking through localStorage and checking the data in the console / Application tab of devtools. EVEN IF they do, YYYY-MM-DD is still easy to understand.