update diagram representing data structure of localstorage by adding a projectId attribute to tasks
this means avoiding having to pass through an additional argument into app controller that pulls the current project from the dom, making the parameters clearer that it is the new task objects' attributes. it also avoids the previous way in the console version where there is a helper that scans through all of the projects to find the one with the task. if there are 100 projects, this is really inefficient.