at the appController level, if you're adding a task, you pass in the project Id and not the project name
project id is more unique. this is to future proof the very likely scenario when a user wants to change a project name. i feel this is the most future-proof way by using the most unique key for a project. from the DOM, we can always get the project ID too using data attributes, which should be coming in the next few commits