do not submit the form in the modal that display details of task. instead, put event listener on the button for save
this is because i can't pass both an event ('e') parameter to stop default action of submitting the form, which isn't the form, AS WELL AS the task object that the modal is referring to. i guess the only other way is putting the task's id on the <form> element as a data-attribute, so that i only need the 'e' parameter (and do like e.target.dataset.id or something). BUT, that would mean i need to retrieve the task, reconstruct it (both done in method in appController) when modifying a task... on the balance and thinking out loud, it just feels more cumbersome, but is an option in the future if i want to align the behaviour of the submit/save buttons on the other modals. the other downside is that i need a separate event listener on the enter key to do the same thing as pressing save, which i'll bear for now