Modify existing tasks' attributes, including project (#24)
Completion _can_ be edited once you open up the task's details modal (and indeed updates localStorage), but not yet from the checkbox in the <main> element of the page - the task <article> elements. Styling is only partially implemented. These further tasks relating to completion of a task should be mopped up in #15 

* create function in separate module called modifyTask.js, which is an event handler function that gets called when the form in an task details modal is submitted

* prevent default behaviour of submitting

* pass in task object to specify

* begin writing modification of tasks at the app controller level

technically this is correct, because we're not trying to manipulate the DOM just yet

* ensure tasks are rebuilt to include methods in original factory at the appController level

* write check to see if project name has changed. if it hasn't, just get all the stored tasks and filter out the current task from it

project name changing means that we need to touch two projects. if not, then we don't even need to touch the projects, because projects only have a reference to the tasks via the tasks' ids!

* add setter for name of a task

keeps it private, rather than exposing the .name... that would render the getter useless!

* use setters to update task attributes

* store task once modified. commented out project HasChanged check because it was wrong

* update diagram representing data structure of localstorage by adding a projectId attribute to tasks

this means avoiding having to pass through an additional argument into app controller that pulls the current project from the dom, making the parameters clearer that it is the new task objects' attributes. it also avoids the previous way in the console version where there is a helper that scans through all of the projects to find the one with the task. if there are 100 projects, this is really inefficient.

* add mappedProjectId attribute to task object factory and add comment

at the moment, a task must have a project (it can't be projectless). a task cannot have more than one project. by default, the task goes to the 'Inbox', ready to be categorised into a project or not at all

* add projectId to viewDetails() method of task so it is stored in localStorage. have a method called 'getProjectObject()' that returns a reconstructed project object

this aligns with how the project also has a method to reconstruct all of its tasks from the array of task ids in each project. ?? question: by importing project and task methods into each other, does this make it too tightly coupled??

* simplify retrieveProjectNames localStorage helper into a single line (shorthand) arrow function

* have localStorage helper to retrieve all project Ids

* begin writing setter for project id. return a console.log statement if the project id does not exist

* complete setter for project id and expose in return of task object

* ensure project id is also passed through and set when task is being recreated from JSON

* hide project id from the task divs

* at the appController level, if you're adding a task, you pass in the project Id and not the project name

project id is more unique. this is to future proof the very likely scenario when a user wants to change a project name. i feel this is the most future-proof way by using the most unique key for a project. from the DOM, we can always get the project ID too using data attributes, which should be coming in the next few commits

* remove unused imports

* instantiate app controller on taskModalsHandler instead of reaching deeper into the layer that is localStorage handlers

* update variable name for clarity. restore recreating project at app controller level

* properly call createProject for inbox and actually give the inbox a name

* make sure that the 'value' of the <option> for dropdown for projects is the projectID and not the projectName

* remove unused console log

* write getter for task's project id (rather than the project object) so that this can be called efficiently in displaying an existing task's project in the dropdown

* remove unused comment

* do not submit the form in the modal that display details of task. instead, put event listener on the button for save

this is because i can't pass both an event ('e') parameter to stop default action of submitting the form, which isn't the form, AS WELL AS the task object that the modal is referring to. i guess the only other way is putting the task's id on the <form> element as a data-attribute, so that i only need the 'e' parameter (and do like e.target.dataset.id or something). BUT, that would mean i need to retrieve the task, reconstruct it (both done in method in appController) when modifying a task... on the balance and thinking out loud, it just feels more cumbersome, but is an option in the future if i want to align the behaviour of the submit/save buttons on the other modals. the other downside is that i need a separate event listener on the enter key to do the same thing as pressing save, which i'll bear for now

* gather all required 'updated' task details in an object from DOM in modifyTask.js

* test whether project has changed once save button is clicked - success!

* test whether project has changed once save button is clicked - success!

* close the modal after submission, and refresh the current project page

* add comments as next steps on how to handle when project has changed

* add method to project object that removes task

otherwise, we'd have to replace task with an array that has been filtered anyway in app controller. better to just do it at the source tbh

* if project has changed, remove the id of the task from the old project and store the project, then add the task id to the new project and store the project, then set the project id to the new project id

* remove unused comment

this will be managed through github issues

* pressing enter key on keyboard on task details modal open is the same as clicking the save button

* remove unused comment

* enforce having inbox first if inbox is not excluded